Make subworks sub-commands
